Broke up and refactored the Premium tab.
