+++ .gitignore
@@ -37,3 +37,4 @@ yarn-error.log*
 /localdata/
 /venv/
 /embeddings_backup.npz
+/AGENTS.md
