+++ app/sitemap.ts
@@ -18,7 +18,13 @@ export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
       priority: 0.9,
     },
     {
-      url: `${baseUrl}/premium`,
+      url: `${baseUrl}/llm-chat`,
+      lastModified: new Date(),
+      changeFrequency: 'weekly',
+      priority: 0.8,
+    },
+    {
+      url: `${baseUrl}/overview-report`,
       lastModified: new Date(),
       changeFrequency: 'weekly',
       priority: 0.8,
